requirements.txt
# ============================================================
# SQL-NL Pipeline — Full Requirements
# Covers: SQL generation, NL rendering, perturbations,
#         LLM experiments (API + vLLM), evaluation, plotting
# ============================================================

# --- Core Pipeline (01-03: SQL gen, NL rendering, perturbations) ---
sqlglot>=28.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# --- LLM API Adapters (04, run_experiments) ---
google-genai>=1.0.0          # Gemini adapter
openai>=1.0.0                # OpenAI adapter
anthropic>=0.20.0            # Anthropic adapter

# --- Local Model Inference ---
vllm>=0.4.0                  # vLLM adapter (installs torch automatically)
bitsandbytes>=0.46.1

# --- Analysis & Plotting ---
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.60.0
